Using standard linux data types.

+++ crc16.c
@@ -71,7 +71,7 @@ void crc16_ccitt_update(unsigned short *crc, unsigned int data)
 }
 
 
-unsigned short crc16_ccitt_block(const unsigned int *buf, int len)
+unsigned short crc16_ccitt_block(const uint8_t *buf, int len)
 {
 	unsigned short crc = CRC16_CCITT_INIT_VALUE;
 	int i;
